Corrected bug on camera introduced when changing the pixel type to float. Also removed BufferedPixelDrawer because it was not addapted to float pixels.

+++ src/scene/Camera.cpp
@@ -10,7 +10,7 @@ void Camera::init() {
 	ze = glm::normalize(ze);
 
 	h = 2 * df * tan(fovy / 2);
-	w = (_resX / _resY) * h;
+	w = ((float)_resX / (float)_resY) * h;
 
 	xe = glm::cross(up, ze);
 	xe = glm::normalize(xe);
@@ -20,7 +20,7 @@ void Camera::init() {
 
 Ray Camera::getPrimaryRay(float x, float y) {
 	Ray ray;
-	glm::vec3 direction = -df * ze + h * ((y / (_resY - 1.0f)) - 0.5f) * ye + w * ((x / (_resX - 1.0f)) - 0.5f) * xe;
+	glm::vec3 direction = -df * ze + h * ((y / ((float)_resY - 1.0f)) - 0.5f) * ye + w * ((x / ((float)_resX - 1.0f)) - 0.5f) * xe;
 	ray.point = eye;
 	ray.direction = glm::normalize(direction);
 	return ray;
